+++ ChaCha/Helper/ByteUtil.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 
-namespace Cryptool.Plugins.ChaCha.Util
+namespace Cryptool.Plugins.ChaCha.Helper
 {
     /// <summary>
     /// This class handles all byte operations.
